Seek to current file position to clear EOF status on Mac OS X.
Mac OS X and Linux differ in behaviour when reading from a file that has previously reached EOF.  On
Linux, any new data that has been appended to the file will be returned.  On Mac OS X, the empty string
will always be returned.  Seeking to the current position in the file resets the EOF flag on Mac OS X
and will allow future reads to work as intended.