#227:scheduler-property-in-build.patch
Patch by Dossy Shiobara <dossy@panoptic.com>: It is useful to know
the name of the scheduler that initiated a build request, i.e., to
conditionally FileUpload the results of a Nightly scheduled build.
This patch passes the Scheduler along the BuildRequest chain so
that Build can setProperty("scheduler", self.scheduler.name) in
Build.setupStatus().