don't retire a missing-notification-timer until you finish accepting a slave
Previously, we retired the timer when we began accepting the slave.  However, if we then detected duplicate connection and rejected one of the connections, then the resulting detach would start a new missing-notification-timer.  With this patch, the missing-notification-timer that is started by detaching a connection is soon retired by the other connection leading to acceptance of the buildslave.  So this probably fixes the other way that issue #211 was happening (after the previous patch that was intended to fix #211).