Don't require a list of builders for buildbot try.
This maintains the current restrictions around builder lists that
prevent one from trying a build that isn't in the list, but allows the
user to delegate the selection to the server by not listing the
builders at all.

I want my users to always try their builds on every build
configuration, but I don't want to be sending out buildbot options all
the time.