Allow users to specify the remote git branch.
This allows for a case where someone has a repository that tracks
someone else's repository, has arbitrary local branches, but wants to
run tries with the delta from the reference repository (i.e. the one
the master knows about) to the local changes.

Without this, it's likely the reference repository will not have the
necessary objects to pull down a base revision to be able to apply
patches for the try to succeed.

This also ensures that the current client's view of the reference
repository is honored.  That is, if the reference repository has moved
forward, the trier's current tip of the remote is used to compute the
delta, and that's sent along as the baserev.