Added ping method to stats builders.
I had a lot of network instabilities that would cause my status client
to believe it was connected and functioning properly when it wasn't.
I started sending pings across periodically and flagging the
connection on the (error) result.  That wasn't a big deal, but it ends
up dumping a bunch of stuff into logs.

This makes a simple ping possible without error.