contrib/git_buildbot.py: Exit normally on --help
OptionParser will print out help text and exit with status 0 if -h or
--help is passed in.  git_buildbot.py was catching all exception in the
option parser, and reporting the exception as an error, which caught
the SystemExit exception that OptionParser threw.  This patch simply
passes SystemExit exception up to be handled normally, so that --help
won't print extra exception information and terminate abnormally.