contrib/git_buildbot.py: Accept a --master paramater
git_buildbot.py had a hard-coded master that it would report to.  This
meant that if there were multiple git repos on a server that needed to
report to different buildbot masters, they would each need a separate
copy of git_buildbot.py.  This patch adds a --master (or -m) parameter
for setting the buildbot master to talk to, which can be passed in
from the hooks/post-receive script in each git repo.