Removing useless methods
At first, I saw that the changes list was being sorted by the revisions, and thought that because of that
it will be necessary to allow sorting by time for the other VCSs. But, then I realized that the sorting
was being done simply so that duplicate changes with same revision will be grouped for removing. This works
no matter if the revision is a number or not.