remove test_limitlogs completely
An strace of this test reveals that its sub-processes extend past the
end of the test, and the resulting SIGCHLDs confuse those subsequent
tests.  This minimal addition to buildbot.tac has been fairly well
tested now, so I'm happy to assume it works, given the trouble this
test has caused.