schedulerdb 3-of-5: BuildRequest/BuildSet
The canonical storage of a build request is now in the DB, but a separate
BuildRequest class is defined (in buildbot/buildrequest.py, split out from
buildbot/process/base.py) to pass around to code that needs it. Likewise the
old BuildSet class has been removed and its funcionality moved into the
database. The scheduling methods on BuildSet (like waitUntilSuccess and
start) have been subsumed by other changes to the Schedulers and Builders, so
buildbot/buildset.py was removed altogether.