Implemented clean master shutdown.
Has a "Shutdown" button on the root web status page, and a new flag for
the authorization object, "cleanShutdown", that controls if users are
allowed to shut down the master or not.

The master will stop assigning builds to slaves.  Once all the slaves
have finished builds, the master process will terminate.

You are allowed to cancel a shutdown in progress, in which case builds
will resume being assigned to slaves, and the master process will no
longer terminate when idle.

The shutdown can also be initiated via manhole by calling
status.cleanShutdown() and cancelled by status.cancelCleanShutdown()