Calculate 'now' for scheduler loop ahead of time.
When the master is over-loaded, it can take long enough between calling
the schedulers and evaluating the results that the result from the
scheduler's run() method is in the past.

By calculating 'now' before running the scheduler's run method, the
assertion that the scheduler's result is > 'now' should be true for
well-behaved schedulers.