Add a GPLv2 copyright header to each python file
This makes the licensing conditions of Buildbot clear from analysis of
any particular file, even if it becomes separated from the package as a
whole.

Files in master/contrib and slave/contrib are not included.

Please note that these changes are made based on my best understanding
of the license and copyright under which each file falls, but that my
analysis may be imperfect.

Fixes #1698