use self.patch in tests, but not in <9.0.0 on py-2.7
The addCleanup method, which self.patch uses to clean up, does not work
correctly on py27.  This was handled in the slave monkeypatch, and is
now handled on the master, too.  This required some reworking of the
gitpoller tests, which are using an unqualified except (in the form of
an errback without f.trap) and thus eat the SkipTest exception.