Pass system and basedir from slaves remote_getSlaveInfo.
'system': This is python's os.name. It is intended to be used to do path manipulation
	on the master that is correct for the slave. We pass this, rather than the path
	separator, to avoid hard-coding that '\' is tied to windows style root-directories.
'basedir': This is the slave base directory, where builder directories are created. This
	is to allow the master to create absolute directories for the slave, for build steps
	that need it.

Right now, there is no master side support for accessing these. This is the minimal support
need slave side, so the 8.3 slave can be used with later masters that use this.

Signed-off-by: Tom Prince <tom.prince@ualberta.net>