Use new-style Python classes.
One change was required to util/init.py in ComparableMixin,
as there was an assumption that types for class instances were always the
same. This isn't true with new-style classes and caused a testsuite
regression. The following cmp of class.name will catch any differences
between the two, so the check was a no-longer useful optimization.