(#184) Blocker: make timeout handling a little tighter.
Add a flag so we know when the Blocker timed out, and return early
from _upstreamStepFinished() in that case. If we timeout, we don't
care about the fate of any upstream steps; we just need to clean them
out of Blocker's data structures and get out of the way.

Also add a test case for timeouts.