Make build object available during substantiation
When substantiate happens a build has already been allocated to
the future slave. It would be awesome to be able to customize
the VM libvirt is substantiating based on build properties.

This patch makes it so.

Now I can control the memory, cpu allocation, and base image
for my latent slaves from properties.