Builder configuration should default mergeRequests to None, not True
Without this change, mergeRequests is True regardless of the global
setting, unless specifically set to something else in the BuilderConfig.

This also fixes an error in _getMergeRequestFn, which was looking at
self.master.mergeRequests, not self.botmaster.mergeRequests, and updates
the tests to match and to have somewhat more readable inputs.

Fixes #2008.