added commandline tool to manually manage users
This is available as `buildbot user` and is
structured very similar to `buildbot sendchange`

This allows users to be added, updated, removed,
and shown via the commandline. The structure of
the manager will allow for more user tools to
manually manage them in the future.

Tests and documentation should explain the new
commandline tool concisely.