added Users page to WebStatus to display users
This adds a page to display users that are currently
in the database. The base page has a list of users
by uid and identifier. Each user has their own page
containing all of the attributes they have, which is
clickable through the user's uids on the base page.

To list all the users in the database, getUsers was
added to db.users to fetch all the uid/identifier pairs
currently stored. This also introduces `showUsersPage`
as an authz option to control who can view the users.