added UsersAuth IAuth instance
This adds a third IAuth instance, in addition to BasicAuth
and HTPasswdAuth, for web authentication. The username and
password pairs are stored in the users table, with the
passwords hashed and salted.

This meant additional bb_username and bb_password columns
in the users table, so the db connector, migrate script,
and command line tool have been updated.

Documentation where appropriate has been updated, giving
information about bb_username/bb_password and UsersAuth.