triggereable: update triggering step with URL for trigged build
When having a lot of parallel build, its difficult to find which
test build is triggered by a compile build.

The triggerable now takes another optional argument which is the
step status that where add the URLforthing(buildstatus) for the
triggered build.

We also remove issue with builder advertising a new build a little
bit early leading the master status not to inform subscribers.

Signed-off-by: Pierre Tardy <pierre.tardy@intel.com>