buildslave/git: Don't fail the whole build if git branch -M can't rename master.
Previous versions of Git allowed the current branch to be renamed. For
some reason, Git 1.7.7 does not allow this. All attempts to rename the
current branch fail with "Cannot force update the current branch." So,
if the rename fails, don't abort.

As part of this, add support for commands that can fail.