authz: rework for cookie based session management
Instead of requiring the user to type password in each and
every form inside buildbot, we implement a simple session
manager.

Existing IAuth interface is augmented with getUserInfos()
which is used to display info about the user
/login and /logout ActionResources are added to manage login and logout

template helpers like advertiseAction() now needs to pass request, so
that authz can get the session cookie from headers

now that we can identify between auth fail and authz fail,
we also add an authz failed page in order to let the user know
whether the authentication failed of the authorization failed.

These changes happen behind the scene, there is no update in the
configuration API, and the user should just notice the change in the
UI (the login form changed)

Signed-off-by: Pierre Tardy <pierre.tardy@intel.com>