Refactor configuration handling
This moves configuration loading and checking out of the master, and
into a new module.  Where possible, configuration is accessed directly
in as many cases as possible, rather than being disseminated across the
master at reconfiguration time.  For services that need to modify their
behavior when reconfigured, the ReconfigurableServiceMixin defines a
general method of notification, eliminating service-specific code from
the master.

Schedulers are still stopped and re-started at every reconfiguration;
this may be changed in the future.