Fix master state tracking.
In the previous version, the master tracked its last_processed_change
using a name/class of 'master'/'buildbot.master.BuildMaster', but
tracked build requests using the master name with class 'BuildMaster'.

The migration script given here combines these two objects into one,
discarding the last_processed_change value if it is ambiguous which
master owns it.  In the end, the object name is the master name, and the
class is 'buildbot.master.BuildMaster'.

This also adds a framework for unit-testing migration scripts, and uses
it.