Change MailNotifier mode to allow for more flexibility
The attached patch changes mode to be a list allowing for example to
receive emails when there is a change and for any successive failures or
only for warnings.

The change also handles backward compatibility when passing a string and
in particular for "all" and "warnings" string modes.

The only change wrt to the current behaviour is that mode="all" will
report whether the build failed, passed or warnings were found instead
of "The Buildbot has finished a build" in the default message. I believe
this is an improvement, ymmv.