Fixes #1799
This patch escapes all pipes in list commands that do not occur by
themself.  This addresses the problem passing in Visual Studio
configuration names, which in their commands are by default patterned
like [...,"Release|Win32",...], which can't be parsed by the shell.

However, this does not touch two different conditions: lone pipes
in a command list (e.g., ['a','|','b'] assumes you want to pipe
the output of a into b) and pipes inside string commands (e.g.,
where the full command is "a | b", not in a list, it is assumed
that you don't want an escaped pipe).

Additionally, this commit adds a skip decorator for the platform,
to clean up that syntax.