Disable checkpoint_fullfsync on all SQLite connections.
The checkpoint_fullfsync pragma is enabled by default on the version of SQLite that
ships with OS X Lion and newer. This setting enables use of the F_FULLSYNC fcntl rather
than fsync when transactions are moved from the WAL back in to the database proper. The
fcntl waits until the disk drive reports that all buffered data has been written to
permanent storage, in constrast to fsync which merely waits until the data has reached
the disk controller. F_FULLSYNC is more robust in the face of catastrophic system failure
but incurs a substantial performance hit for being so.

This change reduces the time it takes to run the buildbot.test.integration.test_upgrade
suite of tests on one machine from over 27 seconds to under 7 seconds.

Fixes #2256.