Fix tests after 82430c when using older versions SQLAlchemy.
Use an object that implements the PoolListener interface to listen
for the connect event when using SQLAlchemy prior to 0.7.

This also moves the registration of the event listener ahead of the
changing of the journal mode to ensure that the database connection
that SQLAlchemy spins up when executing the journal mode pragma also
triggers the event listener, and doesn't miss out on having
checkpoint_fullfsync disabled.