Ensure changes and requests are sorted chronologically
When creating a new SourceStamp, comprising changes are not guaranteed
to come sorted chronologically. Sort them to ensure the most recent
Change's attributes are used for the new SourceStamp.

Also, when merging requests according to the user-selected one with
nextBuild(), keep the qualifying requests in the same relative order
they were in the original superset passed in before the merge.

Change-Id: I95d02b5fae28fa87aa2cd4f2f8e3578317cf2b72