www_haml: allow use of haml for templating
Haml is a nice and simple templating langage, that avoids to write tons of
markups, and has basic facility for loops, conditionnals, and macroing.

Templates are fetched from the static server, compiled into a javascript
function, cached, and executed on demand, with widget as the context.

widgets can implement loadMoreContext() in order to load more context
from the json api

Signed-off-by: Pierre Tardy <tardyp@gmail.com>