Ensure that locks are acquired in FIFO order
To prevent starvation, locks should be acquired in the order in which
they were requested. Particularly, counting locks requested after
exclusive locks must wait until after the exclusive lock is acquired
and subsequently released, even if the owner(s) at the time the lock
is requested are counting and below maxCount.

The test acquires a counting lock, and then starts a build needing an
exclusive lock followed by a build needing a counting lock. The first
counting lock is then released, and the test checks that the exclusive
lock was claimed before the second counting lock.