Make previous build chioce configurable
When prioritizing build requests in an order other than arrival time,
builds' relative order might not always reflect the topological order of
the changes. This change makes the build order to be configurable such
that status transition modes are only sent out when a more recent
*change* affects the status of a less recent one, as opposed to relying
on the default chronological ordering of the builds.

Change-Id: I2009a8b1f4d093a23ef6eea869567aa54b7987b6