slavebuilder.py: add a timeout to ping
On rare conditions, buildslave filesystem configured on NFS can
hang undefinitively. Each process accessing the nfs will hang, thus
the slave will never answer the ping. The slave is not completly dead
and tcp connection typically stays alive. Only layer 5 is blocked.

As the slavebuilder's ping is called with botmaster's activity lock
held, this situation leads to a complete hang of the master's build queue

This fix puts a timeout to the answer of the ping.

Change-Id: I938a0f4365a610bdda1f68f1b298c68977d39d09
Signed-off-by: Alberdi, IonX <ionx.alberdi@intel.com>
Signed-off-by: Pierre Tardy <pierre.tardy@intel.com>