PyFlakes: fix handling of warnings
The buildstep tried to return WARNINGS or FAILURE depending on the
severity of the warnings issued. However, this didn't work since
PyFlakes always return 1 if it finds any warnings or errors. A return
code of 1 is now decoded to mean WARNINGS, and the evaluateCommand
method will then classify the result further.