ArtifactStorage management design doc
Buildbot lacks a framework for storing the build artifacts.
Here is my proposal for building such.

The overall goal is to be able to write backend plugins for
more complex artifact storage systems like artifactory, or apache's
archiva

I volunter for implementing it once we have an agreement on the API,
I plan to implement a simple dictionary of storageManager, in process.build.Build class that Create*ArtifactStorage will populate, and that the BuildStep artifact APIs will proxy.

Signed-off-by: Pierre Tardy <pierre.tardy@intel.com>