Update master/buildbot/status/mail.py
Buildsets do not report correctly via the mail notifier. The 'builder' variable is only referenced by the lambda function, causing the wrong builder to be used during the call to _gotBuilds(). Creating a local copy in the lambda function via default argument was the quickest and cleanest way to make the value stick for each iteration through the loop. This could be pushed forward to 8.7, etc. if accepted.