Added a default for view in the web.authz module.
I added a default to true for the view in the master.status.web.authz
module. If there is no view specified in the master config then the
master will behave the same as it did before view was added. If view is
explicitly set in the master.cfg that argument will be used.