Added the url print when buildbot try exists
I added a feature that prints out the urls associated with that
instance of buildbot try. It only prints when you use the --wait
option. To use you use buildbot try just as you would have before. This
just adds more to what is displayed when the build is completed.