Added --url option to try client
Added a --url option to the try client. When specified the client
will print out the url of the related builds and exit. If wait is
specified then buildbot will print out the url at the end of wait.
When --wait is not specified we need to wait until the build has
started before we can get the url for it.