Fixed the url feature to be more effecient
Changed the way that we implement the url feature. With this
implementation we get the url from the build before it starts to print
out the status messages. Rather than at the end. It simplifies the code
greatly and make is less prone to issue regarding multiple builds from a
single buildrequest.