Add BuildslaveManager, hide protocols behind it
This begins to build an abstraction for protocols, so that buildslaves
and other parts of Buildbot will not know or care if they are using AMP
or PB or something else.

The tricky bit is that most protocols will involve a single listener
that the buildslave manager can handle directly, while PB connections
are shared by the PBManager.  For the moment, I've kept the idea of
"registrations" in place, but we should refactor this eventually.