fix anomalous-backslash-in-string (W1401) pylint warning
This warning is often triggered by regular expressions, and prefixing
to the open quotation mark will make python leave any escape sequences
alone, e.g.:

 - re.compile("\w+")
 + re.compile(r"\w+")

This solution is not specific for regular expressions, it can be applied
to any string that contains backslashes that python should not interpret.

This commit also enables the warning in common/pylintrc.