Git: Fix initial clone with gerrit changes
git clone --branch only works with refs under refs/heads/* and
refs/tags/*, not arbitrary refs. But the Gerrit step will give us a
ref of the form refs/changes/*, since that is where gerrit stores the
submitted changes for us to build.

There appears to be no way to get an arbitrary ref with 'git clone',
as you can with --branch; at least, none I can see. So instead, just
do a regular clone, and run the normal _fetch routine afterwards,
which will fetch the relevant ref and check it out for us.

This bug is mentioned in bug #2485, though #2485 itself is a different
bug.