GerritStatusPush: Fix labels with gerrit >=2.6
Fixes #2550. Gerrit versions 2.6 and newer do not accept the
--verified option if the Verified label does not exist in All-Projects
in the gerrit configuration. Instead, we must give the --label option.
So, use --label for --verified and --code-review so we can work with
all gerrit installations.

While the --code-review option does still work, so does the --label
approach. It seems more likely that --label is more likely to work in
the future than --code-review, so we'll just go with the safer route.

In order to still support older gerrit installations (which do not
support the --label option), query the gerrit installation version
whenever we submit a review. Cache this version query for 10 minutes,
so we do not ask the gerrit server every single time for its version.