event enhancements
This adds more coherency between the event system and the data system, so that
the events are more coherent with the rest api.
Eventually, the goal is that the events are always the same structure as the data for a same resource

This simplifies a lot the system, and make the client much more easy
This comes at a efficieny price. With this design, events will contain more data.

messages are anyway very small, and probably negligible compared to the network load of logs, and slave communication

Signed-off-by: Pierre Tardy <pierre.tardy@intel.com>