Propagate interrupts to remote commands
Currently, source control commands will hang indefinitely if the slave
goes away while while a job is running on it.  You can also see this
work correctly currently by trying to stop a build while a source
control step is running; the command that was currently running will
continue to run until it is completed, and then the build is stopped.

This change does the same thing that `LoggingBuildstep` does; it stores
the currently running command in `self.cmd` while allows `interrupt` to
be called on the `RemoteCommand`.

The `Repo` step actually had fixed this for it, so I've removed that
fix and made it apply to all steps that are based on `Source`.