status_json: Allow double escaping for resources in select
This patch provides a workaround for an issue that would cause queries
of the form

"select=builders/my%20slashed%2Fbuilder"

to be interpreted as a hierarchy of separate resources:

builders
 - my slashed
   - builder

Instead of the intended:

builders
 - my slashed/builder

This way, we can double escape resources with specially treated
characters in them (such as slashes), and still retrieve them properly.

For instance, the query above becomes:

"select=builders/my%2520slashed%252Fbuilder"

instead, and renders the desired result.

Signed-off-by: Anton Lofgren <alofgren@op5.com>