Seed the `getChange` cache from `addChange`.
A common flow is to add a change to database via `addChange`
then immediately read it back using `addChange`, which would
cause a DB access to insert it and another to read it back.
This change seeds the `getChange` cache directly from the
`addChange` call to avoid this second DB access.