change imports to avoid shadowing
in
```
from buildbot.util import poll

class Foo:
    @poll.method
    def someMethod(self):
        pass
    def poll(self):
        pass
```

The `poll` in `@poll.method` correctly refers to the module.  However,
if the order of the methods is reversed, it refers instead to the `poll`
method in the class scope.

Pyflakes only detects this in the problematic ordering, while pylint
detects it even in the working order.  Pylint is probably right -
re-ordering methods shouldn't change behavior!