Added pushing data to the autobahn server
This will use a hybrid PUSH/PULL service, so that we don't need to completely re-write the front-end to deal with smaller parts of the information coming in one at a time.
Once changes have been detected Katana will poke the relevant autobahn server telling it that a specific event has happened with a series of properties.
The autobahn server will then check its registered URLs looking to see if the filters match the properties given, if they match then the URL will be allowed to poll.

Note: The PUSH URLs are still subject to the same poll interval as any of the POLLING ones on top of the check for the PUSH of data

- Added PUSH filters for the global JSON and the builders JSON
- Reduced the allowed updates on the front-end down to 500ms so that we can start doing things a little more instant if we desire