unit tests: handle double fork on upstart system
Don't assume that when the parent process dies, the child's
ppid will become 1.

This is for example not the case when running unit tests on an
upstart based system (e.g. Ubuntu 13.10) from GUI session.

When a user loggs in into the GUI, a new 'init' process is started
by upstart. That new 'init' process will inherit any orphanded child
processes.