Add changes for new branches in GitPoller
With the new branches=True argument for GitPoller to watch all
branches, _process_changes() does nothing if a new branch is pushed. The
only time it adds changes to the database is when a new commit is pushed
on an existing branch. This is unfortunate because this means in order
to have buildbot build all git branches, one must first push the branch,
wait for GitPoller to detect it, and then push new commits on the
branch.

http://trac.buildbot.net/ticket/2837