Added a cache for latest build for builders
This cache will keep information about the latest build on builders with any branch information given,
it will also cache when no build is found and is kept over restarts. This should dramatically reduce the
load on the server when starting up and should allow us to increase the amount of pickles we can keep.

The cache is kept for 30 days by default and can be changed in the config.

- Fixed a bug where the builder was not being saved to disk